requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
openai==1.66.5
scikit-learn==1.4.0
numpy==1.26.4
supabase==2.3.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
